requirements.txt
wheel
absl-py==1.0.0
antlr4-python3-runtime==4.8
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.6.0
google-auth-oauthlib==0.4.6
google-auth==2.6.2
grpcio==1.44.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
markdown==3.3.6
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
pillow==9.0.1
protobuf==3.19.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyyaml==6.0
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8
six==1.16.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.0
#torch==1.11.0
#torchaudio==0.11.0
#torchvision==0.12.0
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
werkzeug==2.0.3
zipp==3.7.0